Implement pareto optimization in extraction
