+++ src/extract/beam_pareto.rs
@@ -1,6 +1,5 @@
-//! `extract::beam_pareto` implements a Pareto-optimal beam search extractor
-//! that considers both area and delay costs.
-
+//! `extract::partial` implements a non-ILP-based extractor based on partial
+//! orderings of learned library sets.
 use egg::{Analysis, CostFunction, DidMerge, EGraph, Id, Language, RecExpr};
 use log::debug;
 use std::{
@@ -11,72 +10,266 @@ use std::{
 
 use crate::{
   ast_node::{Arity, AstNode},
+  extract::cost::{AreaCost, DelayCost, LangCost, LangGain},
   learn::LibId,
   teachable::{BindingExpr, Teachable},
-  extract::beam::OptimizationStrategy,
 };
 
-/// A `LibSelAreaDelay` is a selection of library functions, paired with
-/// area and delay cost values for the expression.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LibSelAreaDelay {
-  /// Area cost of the expression without library functions
-  pub area_cost: usize,
-  /// Delay cost of the expression without library functions
-  pub delay_cost: usize,
-  /// Memoized area_cost + sum of lib area costs
-  pub full_area_cost: usize,
-  /// Memoized delay_cost (taking into account critical path)
-  pub full_delay_cost: usize,
-  /// The selected libraries and their costs (area, delay)
-  pub libs: Vec<(LibId, usize, usize)>,
+#[derive(Debug, Clone, Default)]
+struct EmptyAreaCost;
+
+impl<Op> LangCost<Op> for EmptyAreaCost {
+  fn op_cost(&self, _op: &Op, _args: &[usize]) -> usize {
+    0
+  }
 }
 
-impl LibSelAreaDelay {
+// Simple delay cost model for RdgEgg
+#[derive(Debug, Clone, Default)]
+struct EmptyDelayCost;
+
+impl<Op> LangGain<Op> for EmptyDelayCost {
+  fn op_gain(&self, _op: &Op, _args: &[usize]) -> usize {
+    0
+  }
+}
+
+/// A `CostSet` is a set of pairs; each pair contains a set of library
+/// functions paired with the cost of the current expression/eclass
+/// without the lib fns, and the cost of the lib fns themselves.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct CostSet {
+  /// The set of library selections and their associated costs.
+  /// Invariant: sorted in ascending order of `expr_cost`, except during
+  /// pruning, when it's sorted in order of `full_cost`.
+  pub set: Vec<LibSel>,
+}
+
+impl CostSet {
+  /// Creates a `CostSet` corresponding to introducing a nullary operation.
   #[must_use]
-  pub fn intro_op() -> LibSelAreaDelay {
-    LibSelAreaDelay {
-      libs: Vec::new(),
-      area_cost: 1,
-      delay_cost: 1,
-      full_area_cost: 1,
-      full_delay_cost: 1,
+  pub fn intro_op(op_delay: usize) -> CostSet {
+    let mut set = Vec::with_capacity(10);
+    set.push(LibSel::intro_op(op_delay));
+    CostSet { set }
+  }
+
+  /// Crosses over two `CostSet`s.
+  /// This is essentially a Cartesian product between two `CostSet`s (e.g. if
+  /// each `CostSet` corresponds to an argument of a node) such that paired
+  /// `LibSel`s have their libraries combined and costs added.
+  #[must_use]
+  pub fn cross(&self, other: &CostSet, lps: usize) -> CostSet {
+    let mut set = Vec::new();
+
+    for ls1 in &self.set {
+      for ls2 in &other.set {
+        match ls1.combine(ls2, lps) {
+          None => continue,
+          Some(ls) => {
+            if let Err(pos) = set.binary_search(&ls) {
+              set.insert(pos, ls);
+            }
+          }
+        }
+      }
+    }
+
+    CostSet { set }
+  }
+
+  /// Combines two `CostSets` by unioning them together.
+  /// Used for e.g. different `ENodes` of an `EClass`.
+  pub fn combine(&mut self, other: CostSet) {
+    // println!("combine");
+    let mut cix = 0;
+
+    for elem in other.set {
+      while cix < self.set.len() && elem >= self.set[cix] {
+        cix += 1;
+      } // Nadia: can't this be done with a binary search?
+
+      self.set.insert(cix, elem);
+      cix += 1;
     }
   }
 
-  /// Combines two `LibSelAreaDelay`s. Unions the lib sets, adds
-  /// the area costs, and takes max of delay costs.
+  /// Performs trivial partial order reduction: if `CostSet` A contains a
+  /// superset of the libs of another `CostSet` B, and A has a higher
+  /// `expr_cost` than B, remove A.
+  pub fn unify(&mut self) {
+    // println!("unify");
+    let mut i = 0;
+
+    while i < self.set.len() {
+      let mut j = i + 1;
+
+      while j < self.set.len() {
+        let ls1 = &self.set[i];
+        let ls2 = &self.set[j];
+
+        if ls1.is_subset(ls2) {
+          self.set.remove(j);
+        } else {
+          j += 1;
+        }
+      }
+      i += 1;
+    }
+  }
+
+  /// Increments the expr and full cost of every `LibSel` in this `CostSet`.
+  /// This is done if we e.g. cross all the args of a node, then have to add
+  /// the node itself to the cost.
+  pub fn inc_delay(&mut self, delay: usize) {
+    // println!("inc_cost");
+    for ls in &mut self.set {
+      ls.inc_delay(delay);
+    }
+  }
+
   #[must_use]
-  pub fn combine(&self, other: &LibSelAreaDelay, lps: usize) -> Option<LibSelAreaDelay> {
+  pub fn add_lib(&self, lib: LibId, cost: &CostSet, lps: usize) -> CostSet {
+    // println!("add_lib");
+    // To add a lib, we do a modified cross.
+    let mut set = Vec::new();
+
+    for ls1 in &cost.set {
+      if ls1.libs.iter().any(|l| l.0 == lib) {
+        // If this libsel contains the lib we are defining,
+        // we can't use in in the definition.
+        continue;
+      }
+      for ls2 in &self.set {
+        match ls2.add_lib(lib, ls1, lps) {
+          None => continue,
+          Some(ls) => {
+            if let Err(pos) = set.binary_search(&ls) {
+              set.insert(pos, ls);
+            }
+          }
+        }
+      }
+    }
+
+    CostSet { set }
+  }
+
+  /// prune takes care of two different tasks to reduce the number
+  /// of functions in a `LibSel`:
+  ///
+  /// - If we have an lps limit, we remove anything that has more fns than we
+  ///   allow in a `LibSel`
+  /// - Then, if we still have too many `LibSel`s, we prune based on beam size.
+  ///
+  /// Our pruning strategy preserves n `LibSel`s per # of libs, where
+  /// n is the beam size. In other words, we preserve n `LibSel`s with
+  /// 0 libs, n `LibSel`s with 1 lib, etc.
+  pub fn prune(&mut self, n: usize, lps: usize) {
+    use std::cmp::Reverse;
+
+    let old_set = std::mem::take(&mut self.set);
+
+    // First, we create a table from # of libs to a list of LibSels
+    let mut table: HashMap<usize, BinaryHeap<Reverse<LibSelFC>>> =
+      HashMap::new();
+
+    // We then iterate over all of the LibSels in this set
+    for ls in old_set {
+      let num_libs = ls.libs.len();
+
+      // We don't need to do this anymore, because this is happening in cross:
+      // If lps is set, if num_libs > lps, give up immediately
+      // if num_libs > lps {
+      //     panic!("LibSels that are too large should have been filtered out by
+      // cross!"); }
+
+      let h = table.entry(num_libs).or_default();
+
+      h.push(Reverse(LibSelFC(ls)));
+    }
+
+    // From our table, recombine into a sorted vector
+    let mut set = Vec::new();
+    let beams_per_size = std::cmp::max(1, n / lps);
+
+    for (_sz, mut h) in table {
+      // Take the first n items from the heap
+      let mut i = 0;
+      while i < beams_per_size {
+        if let Some(ls) = h.pop() {
+          let ls = ls.0 .0;
+
+          if let Err(pos) = set.binary_search(&ls) {
+            set.insert(pos, ls);
+          }
+
+          i += 1;
+        } else {
+          break;
+        }
+      }
+    }
+
+    self.set = set;
+  }
+
+  pub fn unify2(&mut self) {
+    // println!("unify");
+    let mut i = 0;
+
+    while i < self.set.len() {
+      let mut j = i + 1;
+
+      while j < self.set.len() {
+        let ls1 = &self.set[i];
+        let ls2 = &self.set[j];
+
+        if ls2.is_subset(ls1) {
+          self.set.remove(j);
+        } else {
+          j += 1;
+        }
+      }
+      i += 1;
+    }
+  }
+}
+
+/// A `LibSel` is a selection of library functions, paired with two
+/// corresponding cost values: the cost of the expression without the library
+/// functions, and the cost of the library functions themselves
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+
+pub struct LibSel {
+  pub expr_delay: usize,
+  pub libs: Vec<(LibId, usize)>,
+}
+
+impl LibSel {
+  #[must_use]
+  pub fn intro_op(op_delay: usize) -> LibSel {
+    LibSel {
+      libs: Vec::new(),
+      expr_delay: op_delay,
+    }
+  }
+
+  /// Combines two `LibSel`s. Unions the lib sets, adds
+  /// the expr
+  #[must_use]
+  pub fn combine(&self, other: &LibSel, lps: usize) -> Option<LibSel> {
     let mut res = self.clone();
 
-    for (k, area_v, delay_v) in &other.libs {
-      match res.libs.binary_search_by_key(k, |(id, _, _)| *id) {
+    for (k, v) in &other.libs {
+      match res.libs.binary_search_by_key(k, |(id, _)| *id) {
         Ok(ix) => {
-          // If we already have this lib, use the better costs
-          if area_v < &res.libs[ix].1 {
-            res.full_area_cost -= res.libs[ix].1 - *area_v;
-            res.libs[ix].1 = *area_v;
-          }
-          if delay_v < &res.libs[ix].2 {
-            // Only update delay if it's on the critical path
-            if res.libs[ix].2 == res.full_delay_cost {
-              // Recalculate full delay cost
-              res.full_delay_cost = res.delay_cost.max(
-                res.libs.iter()
-                  .map(|(_, _, d)| *d)
-                  .max()
-                  .unwrap_or(0)
-              );
-            }
-            res.libs[ix].2 = *delay_v;
+          if v < &res.libs[ix].1 {
+            res.libs[ix].1 = *v;
           }
         }
         Err(ix) => {
-          res.full_area_cost += *area_v;
-          res.libs.insert(ix, (*k, *area_v, *delay_v));
-          // Update delay if this new lib is on the critical path
-          res.full_delay_cost = res.full_delay_cost.max(*delay_v);
+          res.libs.insert(ix, (*k, *v));
           if res.libs.len() > lps {
             return None;
           }
@@ -84,13 +277,7 @@ impl LibSelAreaDelay {
       }
     }
 
-    // Update area cost
-    res.area_cost = self.area_cost + other.area_cost;
-    res.full_area_cost += other.area_cost;
-
-    // Update delay cost (critical path)
-    res.delay_cost = self.delay_cost.max(other.delay_cost);
-    res.full_delay_cost = res.full_delay_cost.max(other.full_delay_cost);
+    res.expr_delay = self.expr_delay + other.expr_delay;
 
     Some(res)
   }
@@ -99,48 +286,25 @@ impl LibSelAreaDelay {
   pub fn add_lib(
     &self,
     lib: LibId,
-    cost: &LibSelAreaDelay,
+    cost: &LibSel,
     lps: usize,
-  ) -> Option<LibSelAreaDelay> {
+  ) -> Option<LibSel> {
     let mut res = self.clone();
-    let area_v = cost.area_cost + 1; // +1 for the lib node area
-    let delay_v = cost.delay_cost + 1; // +1 for the lib node delay
-    let mut full_area_cost = res.full_area_cost;
-    let mut full_delay_cost = res.full_delay_cost;
+    let v = cost.expr_delay;
 
     // Add all nested libs that the lib uses, then add the lib itself.
-    for (nested_lib, nested_area, nested_delay) in &cost.libs {
+    for (nested_lib, v) in &cost.libs {
       let nested_lib = *nested_lib;
-      let nested_area = *nested_area;
-      let nested_delay = *nested_delay;
+      let v = *v;
 
-      match res.libs.binary_search_by_key(&nested_lib, |(id, _, _)| *id) {
+      match res.libs.binary_search_by_key(&nested_lib, |(id, _)| *id) {
         Ok(ix) => {
-          if nested_area < res.libs[ix].1 {
-            full_area_cost -= res.libs[ix].1 - nested_area;
-            res.libs[ix].1 = nested_area;
-          }
-          if nested_delay < res.libs[ix].2 {
-            // Only update full delay if this lib was on the critical path
-            if res.libs[ix].2 == full_delay_cost {
-              // Recalculate full delay
-              full_delay_cost = res.delay_cost.max(
-                res.libs.iter()
-                  .filter(|(id, _, _)| *id != nested_lib)
-                  .map(|(_, _, d)| *d)
-                  .max()
-                  .unwrap_or(0)
-                  .max(nested_delay)
-              );
-            }
-            res.libs[ix].2 = nested_delay;
+          if v < res.libs[ix].1 {
+            res.libs[ix].1 = v;
           }
         }
         Err(ix) => {
-          full_area_cost += nested_area;
-          res.libs.insert(ix, (nested_lib, nested_area, nested_delay));
-          // Update delay if this new lib is on the critical path
-          full_delay_cost = full_delay_cost.max(nested_delay);
+          res.libs.insert(ix, (nested_lib, v));
           if res.libs.len() > lps {
             return None;
           }
@@ -148,59 +312,35 @@ impl LibSelAreaDelay {
       }
     }
 
-    match res.libs.binary_search_by_key(&lib, |(id, _, _)| *id) {
+    match res.libs.binary_search_by_key(&lib, |(id, _)| *id) {
       Ok(ix) => {
-        if area_v < res.libs[ix].1 {
-          full_area_cost -= res.libs[ix].1 - area_v;
-          res.libs[ix].1 = area_v;
-        }
-        if delay_v < res.libs[ix].2 {
-          // Only update full delay if this lib was on the critical path
-          if res.libs[ix].2 == full_delay_cost {
-            // Recalculate full delay
-            full_delay_cost = res.delay_cost.max(
-              res.libs.iter()
-                .filter(|(id, _, _)| *id != lib)
-                .map(|(_, _, d)| *d)
-                .max()
-                .unwrap_or(0)
-                .max(delay_v)
-            );
-          }
-          res.libs[ix].2 = delay_v;
+        if v < res.libs[ix].1 {
+          res.libs[ix].1 = v;
         }
       }
       Err(ix) => {
-        full_area_cost += area_v;
-        res.libs.insert(ix, (lib, area_v, delay_v));
-        // Update delay if this new lib is on the critical path
-        full_delay_cost = full_delay_cost.max(delay_v);
+        res.libs.insert(ix, (lib, v));
         if res.libs.len() > lps {
           return None;
         }
       }
     }
 
-    res.full_area_cost = full_area_cost;
-    res.full_delay_cost = full_delay_cost;
     Some(res)
   }
 
-  pub fn inc_cost(&mut self) {
-    self.area_cost += 1;
-    self.full_area_cost += 1;
-    self.delay_cost += 1;
-    self.full_delay_cost += 1;
+  pub fn inc_delay(&mut self, delay: usize) {
+    self.expr_delay += delay;
   }
 
   /// O(n) subset check
   #[must_use]
-  pub fn is_subset(&self, other: &LibSelAreaDelay) -> bool {
+  pub fn is_subset(&self, other: &LibSel) -> bool {
     let mut oix = 0;
 
     // For every element in this LibSel, we want to see
     // if it exists in other.
-    'outer: for (lib, _, _) in &self.libs {
+    'outer: for (lib, _) in &self.libs {
       loop {
         // If oix is beyond the length of other, return false.
         if oix >= other.libs.len() {
@@ -213,12 +353,14 @@ impl LibSelAreaDelay {
             oix += 1;
           }
           Ordering::Equal => {
-            // If other[oix] is equal to lib, continue in the outer loop and increment oix
+            // If other[oix] is equal to lib, continue in the outer loop and
+            // increment oix
             oix += 1;
             continue 'outer;
           }
           Ordering::Greater => {
-            // Otherwise if it's larger, there was no element equal. Not subset, ret false.
+            // Otherwise if it's larger, there was no element equal. Not subset,
+            // ret false.
             return false;
           }
         }
@@ -228,283 +370,103 @@ impl LibSelAreaDelay {
     // We made it! ret true
     true
   }
-
-  /// Calculate the combined cost based on the optimization strategy
-  pub fn combined_cost(&self, strategy: &OptimizationStrategy) -> f64 {
-    match strategy {
-      OptimizationStrategy::MinArea => self.full_area_cost as f64,
-      OptimizationStrategy::MinDelay => self.full_delay_cost as f64,
-      OptimizationStrategy::Balanced(weight) => {
-        // Normalize area and delay to be between 0 and 1
-        // This is a simple approach - in practice you might want to use
-        // more sophisticated normalization based on the range of values
-        let area_norm = self.full_area_cost as f64;
-        let delay_norm = self.full_delay_cost as f64;
-        
-        // Weighted sum
-        (1.0 - weight) * area_norm + *weight * delay_norm
-      }
-    }
-  }
-
-  /// Dominates check for Pareto optimality
-  /// Returns true if self dominates other (self is better in at least one dimension and not worse in any)
-  pub fn dominates(&self, other: &LibSelAreaDelay) -> bool {
-    // Check if self is better or equal in all dimensions
-    let area_better_eq = self.full_area_cost <= other.full_area_cost;
-    let delay_better_eq = self.full_delay_cost <= other.full_delay_cost;
-    
-    // And strictly better in at least one dimension
-    let strictly_better = self.full_area_cost < other.full_area_cost || 
-                          self.full_delay_cost < other.full_delay_cost;
-    
-    area_better_eq && delay_better_eq && strictly_better
-  }
 }
 
-/// A wrapper around `LibSelAreaDelay`s that orders based on combined cost.
-#[derive(Debug, Clone)]
-struct LibSelCombined<'a>(pub(crate) LibSelAreaDelay, &'a OptimizationStrategy);
-
-impl<'a> PartialEq for LibSelCombined<'a> {
-    fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
-    }
-}
-
-impl<'a> Eq for LibSelCombined<'a> {}
-
-impl<'a> PartialOrd for LibSelCombined<'a> {
-  fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-    self.0.combined_cost(self.1).partial_cmp(&other.0.combined_cost(self.1))
-  }
-}
-
-impl<'a> Ord for LibSelCombined<'a> {
-  fn cmp(&self, other: &Self) -> Ordering {
-    self.partial_cmp(other).unwrap_or(Ordering::Equal)
-  }
-}
-
-/// A `CostSetAreaDelay` is a set of pairs; each pair contains a set of library
-/// functions paired with the area and delay costs.
+/// A wrapper around `LibSel`s that orders based on their full cost.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CostSetAreaDelay {
-  /// The set of library selections and their associated costs.
-  pub set: Vec<LibSelAreaDelay>,
-}
+struct LibSelFC(pub(crate) LibSel);
 
-impl CostSetAreaDelay {
-  /// Creates a `CostSetAreaDelay` corresponding to introducing a nullary operation.
-  #[must_use]
-  pub fn intro_op() -> CostSetAreaDelay {
-    let mut set = Vec::with_capacity(10);
-    set.push(LibSelAreaDelay::intro_op());
-    CostSetAreaDelay { set }
-  }
-
-  /// Crosses over two `CostSetAreaDelay`s.
-  #[must_use]
-  pub fn cross(&self, other: &CostSetAreaDelay, lps: usize) -> CostSetAreaDelay {
-    let mut set: Vec<LibSelAreaDelay> = Vec::new();
-
-    for ls1 in &self.set {
-      for ls2 in &other.set {
-        match ls1.combine(ls2, lps) {
-          None => continue,
-          Some(ls) => {
-            // Check if this is dominated by any existing solution
-            let mut is_dominated = false;
-            for existing in &set {
-              if existing.dominates(&ls) {
-                is_dominated = true;
-                break;
-              }
-            }
-            
-            if !is_dominated {
-              // Remove any solutions that this dominates
-              set.retain(|existing| !ls.dominates(existing));
-              set.push(ls);
-            }
-          }
-        }
-      }
+impl PartialOrd for LibSelFC {
+  fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+    let r = self.0.expr_delay.cmp(&other.0.expr_delay);
+    if !r.is_eq() {
+      return Some(r);
     }
 
-    CostSetAreaDelay { set }
-  }
-
-  /// Combines two `CostSetAreaDelay`s by unioning them together.
-  pub fn combine(&mut self, other: CostSetAreaDelay) {
-    for elem in other.set {
-      // Check if this is dominated by any existing solution
-      let mut is_dominated = false;
-      for existing in &self.set {
-        if existing.dominates(&elem) {
-          is_dominated = true;
-          break;
-        }
-      }
-      
-      if !is_dominated {
-        // Remove any solutions that this dominates
-        self.set.retain(|existing| !elem.dominates(existing));
-        self.set.push(elem);
-      }
-    }
-  }
-
-  /// Performs Pareto-optimal unification
-  pub fn unify(&mut self) {
-    let mut i = 0;
-    
-    while i < self.set.len() {
-      let mut j = i + 1;
-      
-      while j < self.set.len() {
-        let ls1 = &self.set[i];
-        let ls2 = &self.set[j];
-        
-        if ls1.dominates(ls2) {
-          self.set.remove(j);
-        } else if ls2.dominates(ls1) {
-          self.set.remove(i);
-          j = i + 1; // Reset j since i was removed
-        } else {
-          j += 1;
-        }
-      }
-      i += 1;
-    }
-  }
-
-  /// Increments the area and delay cost of every `LibSelAreaDelay` in this `CostSetAreaDelay`.
-  pub fn inc_cost(&mut self) {
-    for ls in &mut self.set {
-      ls.inc_cost();
-    }
-  }
-
-  #[must_use]
-  pub fn add_lib(&self, lib: LibId, cost: &CostSetAreaDelay, lps: usize) -> CostSetAreaDelay {
-    let mut set: Vec<LibSelAreaDelay> = Vec::new();
-
-    for ls1 in &cost.set {
-      if ls1.libs.iter().any(|(l, _, _)| l == &lib) {
-        // If this libsel contains the lib we are defining,
-        // we can't use it in the definition.
-        continue;
-      }
-      for ls2 in &self.set {
-        match ls2.add_lib(lib, ls1, lps) {
-          None => continue,
-          Some(ls) => {
-            // Check if this is dominated by any existing solution
-            let mut is_dominated = false;
-            for existing in &set {
-              if existing.dominates(&ls) {
-                is_dominated = true;
-                break;
-              }
-            }
-            
-            if !is_dominated {
-              // Remove any solutions that this dominates
-              set.retain(|existing| !ls.dominates(existing));
-              set.push(ls);
-            }
-          }
-        }
-      }
-    }
-
-    CostSetAreaDelay { set }
-  }
-
-  /// Prune the Pareto front to keep a manageable number of solutions
-  pub fn prune(&mut self, n: usize, lps: usize, strategy: &OptimizationStrategy) {
-    use std::cmp::Reverse;
-
-    if self.set.len() <= n {
-      return; // No need to prune
-    }
-
-    let old_set = std::mem::take(&mut self.set);
-
-    // First, we create a table from # of libs to a list of LibSels
-    let mut table: HashMap<usize, BinaryHeap<Reverse<LibSelCombined>>> =
-      HashMap::new();
-
-    // We then iterate over all of the LibSels in this set
-    for ls in old_set {
-      let num_libs = ls.libs.len();
-      let h = table.entry(num_libs).or_default();
-      h.push(Reverse(LibSelCombined(ls, strategy)));
-    }
-
-    // From our table, recombine into a vector
-    let mut set = Vec::new();
-    let beams_per_size = std::cmp::max(1, n / lps);
-
-    for (_sz, mut h) in table {
-      // Take the first n items from the heap
-      let mut i = 0;
-      while i < beams_per_size {
-        if let Some(ls) = h.pop() {
-          set.push(ls.0.0);
-          i += 1;
-        } else {
-          break;
-        }
-      }
-    }
-
-    self.set = set;
+    Some(self.0.libs.cmp(&other.0.libs))
   }
 }
 
-/// Analysis that uses Pareto-optimal beam search with area and delay costs
-#[derive(Debug, Clone, Copy)]
-pub struct BeamAreaDelay {
-  /// The number of `LibSelAreaDelay`s to keep per `EClass`.
+impl Ord for LibSelFC {
+  fn cmp(&self, other: &Self) -> Ordering {
+    self.partial_cmp(other).unwrap()
+  }
+}
+
+// --------------------------------
+// --- The actual Analysis part ---
+// --------------------------------
+
+use std::marker::PhantomData;
+
+#[derive(Debug, Clone)]
+pub struct PartialLibCost<Op, LA, LD>
+where
+  Op: Clone + Debug + Ord + std::hash::Hash + Teachable + Arity,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
+{
+  /// The number of `LibSel`s to keep per `EClass`.
   beam_size: usize,
   inter_beam: usize,
-  /// The maximum number of libs per lib selection.
+  /// The maximum number of libs per lib selection. Any lib selections with a
+  /// larger amount will be pruned.
   lps: usize,
-  /// The optimization strategy to use
-  strategy: OptimizationStrategy,
+  lang_cost: LA,
+  lang_gain: LD,
+  /// Marker to indicate that this struct uses the Op type parameter
+  phantom: PhantomData<Op>,
 }
 
-impl BeamAreaDelay {
+impl<Op, LA, LD> PartialLibCost<Op, LA, LD>
+where
+  Op: Clone + Debug + Ord + std::hash::Hash + Teachable + Arity,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
+{
   #[must_use]
   pub fn new(
     beam_size: usize,
     inter_beam: usize,
     lps: usize,
-    strategy: OptimizationStrategy,
-  ) -> BeamAreaDelay {
-    BeamAreaDelay { beam_size, inter_beam, lps, strategy }
+    lang_cost: LA,
+    lang_gain: LD,
+  ) -> PartialLibCost<Op, LA, LD> {
+    PartialLibCost {
+      beam_size,
+      inter_beam,
+      lps,
+      lang_cost,
+      lang_gain,
+      phantom: PhantomData,
+    }
   }
 
   #[must_use]
-  pub fn empty() -> BeamAreaDelay {
-    BeamAreaDelay { 
-      beam_size: 0, 
-      inter_beam: 0, 
+  pub fn empty() -> PartialLibCost<Op, LA, LD> {
+    PartialLibCost {
+      beam_size: 0,
+      inter_beam: 0,
       lps: 1,
-      strategy: OptimizationStrategy::Balanced(0.5),
+      lang_cost: LA::default(),
+      lang_gain: LD::default(),
+      phantom: PhantomData,
     }
   }
 }
 
-impl Default for BeamAreaDelay {
+impl<Op, LA, LD> Default for PartialLibCost<Op, LA, LD>
+where
+  Op: Clone + Debug + Ord + std::hash::Hash + Teachable + Arity,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
+{
   fn default() -> Self {
-    BeamAreaDelay::empty()
+    PartialLibCost::empty()
   }
 }
 
-impl<Op> Analysis<AstNode<Op>> for BeamAreaDelay
+impl<Op, LA, LD> Analysis<AstNode<Op>> for PartialLibCost<Op, LA, LD>
 where
   Op: Ord
     + std::hash::Hash
@@ -516,44 +478,63 @@ where
     + Send
     + Sync
     + 'static,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
 {
-  type Data = CostSetAreaDelay;
+  type Data = CostSet;
 
   fn merge(&mut self, to: &mut Self::Data, from: Self::Data) -> DidMerge {
+    // println!("merge");
+    // println!("{:?}", to);
+    // println!("{:?}", &from);
     let a0 = to.clone();
 
-    // Merging consists of combination, followed by unification and beam pruning.
+    // Merging consists of combination, followed by unification and beam
+    // pruning.
     to.combine(from.clone());
     to.unify();
-    to.prune(self.beam_size, self.lps, &self.strategy);
+    to.prune(self.beam_size, self.lps);
 
+    // println!("{:?}", to);
+    // println!("{} {}", &a0 != to, to != &from);
+    // TODO: be more efficient with how we do this
     DidMerge(&a0 != to, to != &from)
+    // DidMerge(false, false)
   }
 
   fn make(
     egraph: &mut EGraph<AstNode<Op>, Self>,
     enode: &AstNode<Op>,
   ) -> Self::Data {
+    // println!("make");
     let x = |i: &Id| &egraph[*i].data;
-    let self_ref = egraph.analysis;
+
+    let self_ref = &egraph.analysis;
 
     match Teachable::as_binding_expr(enode) {
       Some(BindingExpr::Lib(id, f, b)) => {
         // This is a lib binding!
+        // cross e1, e2 and introduce a lib!
         let mut e = x(b).add_lib(id, x(f), self_ref.lps);
         e.unify();
-        e.prune(self_ref.beam_size, self_ref.lps, &self_ref.strategy);
+        e.prune(self_ref.beam_size, self_ref.lps);
         e
       }
       Some(_) | None => {
         // This is some other operation of some kind.
+        // We test the arity of the function
+        let op_delay = self_ref.lang_gain.op_gain(enode.operation(), &[]);
+        // println!("op: {:#?}", enode.operation());
+        // println!("op_delay: {:#?}", op_delay);
+        // println!("number of args: {:#?}", enode.args().len());
         if enode.is_empty() {
           // 0 args. Return intro.
-          CostSetAreaDelay::intro_op()
+
+          CostSet::intro_op(op_delay)
         } else if enode.args().len() == 1 {
           // 1 arg. Get child cost set, inc, and return.
           let mut e = x(&enode.args()[0]).clone();
-          e.inc_cost();
+          e.inc_delay(op_delay);
           e
         } else {
           // 2+ args. Cross/unify time!
@@ -563,64 +544,329 @@ where
             e = e.cross(x(cs), self_ref.lps);
             // Intermediate prune.
             e.unify();
-            e.prune(self_ref.inter_beam, self_ref.lps, &self_ref.strategy);
+            e.prune(self_ref.inter_beam, self_ref.lps);
           }
 
           e.unify();
-          e.prune(self_ref.beam_size, self_ref.lps, &self_ref.strategy);
-          e.inc_cost();
+          e.prune(self_ref.beam_size, self_ref.lps);
+          e.inc_delay(op_delay);
           e
         }
       }
     }
   }
+
+  // For debugging
+  // fn modify(egraph: &mut EGraph<AstNode<Op>, id: Id) {
+  //     println!("merge {}", id);
+  //     println!("{:?}", &egraph[id].data)
+  // }
 }
 
-/// Cost function that considers both area and delay for extraction
-#[derive(Debug, Clone, Copy)]
-pub struct AreaDelayCost<LA, LD> {
-  pub area_cost: LA,
-  pub delay_cost: LD,
-  pub strategy: OptimizationStrategy,
+/// Library context is a set of library function names.
+/// It is used in the extractor to represent the fact that we are extracting
+/// inside (nested) library definitions.
+#[derive(Debug, Clone, PartialEq, Eq, std::hash::Hash)]
+struct LibContext {
+  set: Vec<LibId>,
 }
 
-impl<LA, LD, Op> CostFunction<AstNode<Op>> for AreaDelayCost<LA, LD>
+impl LibContext {
+  fn new() -> Self {
+    Self { set: Vec::new() }
+  }
+
+  /// Add a new lib to the context if not yet present,
+  /// keeping it sorted.
+  fn add(&mut self, lib_id: LibId) {
+    if let Err(pos) = self.set.binary_search(&lib_id) {
+      self.set.insert(pos, lib_id);
+    }
+  }
+
+  /// Does the context contain the given lib?
+  fn contains(&self, lib_id: LibId) -> bool {
+    self.set.binary_search(&lib_id).is_ok()
+  }
+}
+
+type MaybeExpr<Op> = Option<RecExpr<AstNode<Op>>>;
+
+/// This is here for debugging purposes.
+fn display_maybe_expr<
+  Op: Clone + std::fmt::Debug + std::hash::Hash + Ord + std::fmt::Display,
+>(
+  maybe_expr: &MaybeExpr<Op>,
+) -> String {
+  match maybe_expr {
+    Some(expr) => expr.pretty(100),
+    _ => "<none>".to_string(),
+  }
+}
+
+/// Extractor that minimizes AST size but ignores the cost of library
+/// definitions (which will be later lifted to the top).
+/// The main difference between this and a standard extractor is that
+/// instead of finding the best expression *per eclass*,
+/// we need to find the best expression *per eclass and lib context*.
+/// This is because when extracting inside library definitions,
+/// we are not allowed to use those libraries;
+/// so the best expression is different depending on which library defs we are
+/// currently inside.
+#[derive(Debug)]
+pub struct LibExtractor<
+  'a,
+  Op: Clone
+    + std::fmt::Debug
+    + std::hash::Hash
+    + Ord
+    + Teachable
+    + std::fmt::Display,
+  N: Analysis<AstNode<Op>>,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
+> {
+  /// Remembers the best expression so far for each pair of class id and lib
+  /// context; if an entry is absent, we haven't visited this class in this
+  /// context yet; if an entry is `None`, it's currently under processing,
+  /// but we have no results for it yet; if an entry is `Some(_)`, we have
+  /// found an expression for it (but it might still be improved).
+  memo: HashMap<Id, HashMap<LibContext, MaybeExpr<Op>>>,
+  /// Current lib context:
+  /// contains all lib ids inside whose definitions we are currently
+  /// extracting.
+  lib_context: LibContext,
+  /// The egraph to extract from.
+  egraph: &'a EGraph<AstNode<Op>, N>,
+  /// This is here for pretty debug messages.
+  indent: usize,
+  lang_cost: LA,
+  lang_gain: LD,
+  /// The relative weight of area cost and delay cost, from 0.0 (all areaa) to
+  /// 1.0 (all delay)
+  strategy: f32,
+}
+
+impl<'a, Op, N, LA, LD> LibExtractor<'a, Op, N, LA, LD>
 where
-  LA: crate::extract::cost::LangCost<Op> + Clone,
-  LD: crate::extract::cost::LangGain<Op> + Clone,
-  Op: Ord + std::hash::Hash + Debug + Teachable + Clone,
+  Op: Clone
+    + std::fmt::Debug
+    + std::hash::Hash
+    + Ord
+    + Teachable
+    + std::fmt::Display
+    + Arity,
+  N: Analysis<AstNode<Op>> + Clone,
+  LA: LangCost<Op> + Clone + Default,
+  LD: LangGain<Op> + Clone + Default,
 {
-  type Cost = usize;
+  /// Create a lib extractor for the given egraph
+  pub fn new(
+    egraph: &'a EGraph<AstNode<Op>, N>,
+    lang_cost: LA,
+    lang_gain: LD,
+    strategy: f32,
+  ) -> Self {
+    Self {
+      memo: HashMap::new(),
+      lib_context: LibContext::new(),
+      egraph,
+      indent: 0,
+      lang_cost,
+      lang_gain,
+      strategy,
+    }
+  }
 
-  fn cost<C>(&mut self, enode: &AstNode<Op>, mut costs: C) -> Self::Cost
-  where
-    C: FnMut(Id) -> Self::Cost,
-  {
-    match enode.as_binding_expr() {
-      Some(BindingExpr::Lib(_, _, body)) => costs(*body),
-      _ => {
-        // Get costs of children
-        let arg_costs: Vec<usize> = enode.args().iter().map(|&id| costs(id)).collect();
-        
-        // Calculate area cost
-        let area = arg_costs.iter().sum::<usize>() + 
-                  self.area_cost.op_cost(enode.operation(), &arg_costs);
-        
-        // Calculate delay cost
-        let max_child_cost = arg_costs.iter().max().copied().unwrap_or(0);
-        let delay = max_child_cost + 
-                   self.delay_cost.op_gain(enode.operation(), &arg_costs);
-        
-        // Combine costs based on strategy
-        match self.strategy {
-          OptimizationStrategy::MinArea => area,
-          OptimizationStrategy::MinDelay => delay,
-          OptimizationStrategy::Balanced(weight) => {
-            // Simple weighted combination
-            ((1.0 - weight) * area as f64 + weight * delay as f64) as usize
+  /// Get best best expression for `id` in the current lib context.
+  fn get_from_memo(&self, id: Id) -> Option<&MaybeExpr<Op>> {
+    self.memo.get(&id)?.get(&self.lib_context)
+  }
+
+  /// Set best best expression for `id` in the current lib context.
+  fn insert_into_memo(&mut self, id: Id, val: MaybeExpr<Op>) {
+    self
+      .memo
+      .entry(id)
+      .or_default()
+      .insert(self.lib_context.clone(), val);
+  }
+
+  /// Extract the smallest expression for the eclass `id`.
+  /// # Panics
+  /// Panics if extraction fails
+  /// (this should never happen because the e-graph must contain a non-cyclic
+  /// expression)
+  pub fn best(&mut self, id: Id) -> RecExpr<AstNode<Op>> {
+    // Populate the memo:
+    self.extract(id);
+    // println!("id: {:#?}", id);
+    // println!("{:#?}", self.egraph[id]);
+    // Get the best expression from the memo:
+    self.get_from_memo(id).unwrap().clone().unwrap()
+  }
+
+  /// Expression gain used by this extractor
+  pub fn cost(&self, expr: &RecExpr<AstNode<Op>>) -> f32 {
+    let delay_cost = DelayCost::new(self.lang_gain.clone()).cost_rec(expr);
+    let selected_delay_cost = match delay_cost.2 {
+      true => delay_cost.0,
+      false => delay_cost.1,
+    };
+    let area_cost = AreaCost::new(self.lang_cost.clone()).cost_rec(expr);
+    let selected_area_cost = match area_cost.2 {
+      true => area_cost.0,
+      false => area_cost.1,
+    };
+    self.strategy * (selected_delay_cost as f32)
+      + (1.0 - self.strategy) * (selected_area_cost as f32)
+  }
+
+  /// Extract the expression with the largest gain from the eclass id and its
+  /// descendants in the current context, storing results in the memo
+  fn extract(&mut self, id: Id) {
+    self.debug_indented(&format!("extracting eclass {id}"));
+    if let Some(res) = self.get_from_memo(id) {
+      // This node has already been visited in this context (either done or
+      // under processing)
+      self.debug_indented(&format!(
+        "visited, memoized value: {}",
+        display_maybe_expr(res)
+      ));
+    } else {
+      // Initialize memo with None to prevent infinite recursion in case of
+      // cycles in the egraph
+      self.insert_into_memo(id, None);
+      // Extract a candidate expression from each node
+      // println!("Extracting eclass {:#?}", self.egraph[id]);
+      for node in self.egraph[id].iter() {
+        match self.extract_node(node) {
+          None => (), // Extraction for this node failed (must be a cycle)
+          Some(cand) => {
+            // Extraction succeeded: check if cand is better than what we have
+            // so far print!("eclss: {}, ", id);
+            // print!("node: {}, ", node.operation());
+            // print!("cand: {}, ", cand.pretty(100));
+            // println!("cand gain: {}", Self::gain(&self, &cand));
+            match self.get_from_memo(id).unwrap() {
+              // If we already had an expression and it was better, do nothing
+              Some(prev)
+                if Self::cost(&self, prev) <= Self::cost(&self, &cand) =>
+              {
+                ()
+              }
+              // Otherwise, update the memo;
+              // note that updating the memo after each better candidate is
+              // found instead of at the end is slightly
+              // suboptimal (because it might cause us to go around some cycles
+              // once), but the code is simpler and it doesn't
+              // matter too much.
+              _ => {
+                self.debug_indented(&format!(
+                  "new best for {id}: {} (gain {})",
+                  cand.pretty(100),
+                  Self::cost(&self, &cand)
+                ));
+                self.insert_into_memo(id, Some(cand));
+              }
+            }
           }
         }
       }
     }
   }
-} 
+
+  /// Extract the smallest expression from `node`.
+  fn extract_node(&mut self, node: &AstNode<Op>) -> MaybeExpr<Op> {
+    self.debug_indented(&format!("extracting node {node:?}"));
+    if let Some(BindingExpr::Lib(lid, _, _)) = node.as_binding_expr() {
+      if self.lib_context.contains(lid) {
+        // This node is a definition of one of the libs, whose definition we are
+        // currently extracting: do not go down this road since it leads
+        // to lib definitions using themselves
+        self.debug_indented(&format!("encountered banned lib: {lid}"));
+        return None;
+      }
+    }
+    // Otherwise: extract all children
+    let mut child_indexes = vec![];
+    self.extract_children(node, 0, vec![], &mut child_indexes)
+  }
+
+  /// Process the children of `node` starting from index `current`
+  /// and accumulate results in `partial expr`;
+  /// `child_indexes` stores the indexes of already processed children within
+  /// `partial_expr`, so that we can use them in the `AstNode` at the end.
+  fn extract_children(
+    &mut self,
+    node: &AstNode<Op>,
+    current: usize,
+    mut partial_expr: Vec<AstNode<Op>>,
+    child_indexes: &mut Vec<usize>,
+  ) -> MaybeExpr<Op> {
+    if current == node.children().len() {
+      // Done with children: add ourselves to the partial expression and return
+      let child_ids: Vec<Id> =
+        child_indexes.iter().map(|x| (*x).into()).collect();
+      let root = AstNode::new(node.operation().clone(), child_ids);
+      partial_expr.push(root);
+      Some(partial_expr.into())
+    } else {
+      // If this is the first child of a lib node (i.e. lib definition) add this
+      // lib to the context:
+      let old_lib_context = self.lib_context.clone();
+      if let Some(BindingExpr::Lib(lid, _, _)) = node.as_binding_expr() {
+        if current == 0 {
+          self.debug_indented(&format!(
+            "processing first child of {node:?}, adding {lid} to context"
+          ));
+          self.lib_context.add(lid);
+        }
+      }
+
+      // Process the current child
+      let child = &node.children()[current];
+      self.indent += 1;
+      self.extract(*child);
+      self.indent -= 1;
+      // We need to get the result before restoring the context
+      let child_res = self.get_from_memo(*child).unwrap().clone();
+      // Restore lib context
+      self.lib_context = old_lib_context;
+
+      match child_res {
+        None => None, /* Failed to extract a child, so the extraction of */
+        // this node fails
+        Some(expr) => {
+          // We need to clone the expr because we're going to mutate it (offset
+          // child indexes), and we don't want it to affect the memo
+          // result for child.
+          let mut new_expr = expr.as_ref().to_vec();
+          for n in &mut new_expr {
+            // Increment all indexes inside `n` by the current expression
+            // length; this is needed to make a well-formed
+            // `RecExpr`
+            Self::offset_children(n, partial_expr.len());
+          }
+          partial_expr.extend(new_expr);
+          child_indexes.push(partial_expr.len() - 1);
+          self.extract_children(node, current + 1, partial_expr, child_indexes)
+        }
+      }
+    }
+  }
+
+  /// Add `offset` to all children of `node`
+  fn offset_children(node: &mut AstNode<Op>, offset: usize) {
+    for child in node.children_mut() {
+      let child_index: usize = (*child).into();
+      *child = (child_index + offset).into();
+    }
+  }
+
+  /// Print a debug message with the current indentation
+  /// TODO: this should be a macro
+  fn debug_indented(&self, msg: &str) {
+    debug!("{:indent$}{msg}", "", indent = 2 * self.indent);
+  }
+}
