+++ src/lib.rs
@@ -45,7 +45,7 @@ pub use learn::{
   DiscriminantEq, LearnedLibrary, LearnedLibraryBuilder, LibId, ParseLibIdError,
 };
 pub use runner::{
-  BabbleResult, BabbleRunner, BeamConfig, BeamRunner, ParetoConfig, ParetoRunner, KnapsackConfig, KnapsackRunner,
+  BabbleResult, BabbleRunner, BeamConfig, BeamRunner, ParetoConfig, ParetoRunner,
 };
 pub use teachable::{
   BindingExpr, DeBruijnIndex, ParseDeBruijnIndexError, Teachable,
