+++ src/runner.rs
@@ -19,12 +19,11 @@ use log::{debug, info};
 
 use crate::{
   extract::{
-    self, apply_libs, apply_libs_area_delay, apply_libs_knapsack,
+    self, apply_libs,
     apply_libs_pareto,
-    beam::{OptimizationStrategy, PartialLibCost},
-    beam_knapsack,
-    beam_pareto::{BeamAreaDelay, CostSetAreaDelay, LibSelAreaDelay},
-    cost::{DelayCost, LangCost, LangGain},
+    beam::{PartialLibCost},
+    beam_pareto,
+    cost::{AreaCost, DelayCost, LangCost, LangGain},
   },
   Arity, AstNode, COBuilder, DiscriminantEq, Expr, LearnedLibraryBuilder,
   Pretty, Printable, Teachable,
@@ -417,346 +416,8 @@ where
   }
 }
 
-/// Configuration for Pareto-optimal beam search
-#[derive(Debug, Clone, Copy)]
-pub struct ParetoConfig {
-  /// The final beam size to use
-  pub final_beams: usize,
-  /// The inter beam size to use
-  pub inter_beams: usize,
-  /// The number of times to apply library rewrites
-  pub lib_iter_limit: usize,
-  /// The number of libs to learn at a time
-  pub lps: usize,
-  /// Whether to learn "library functions" with no arguments
-  pub learn_constants: bool,
-  /// Maximum arity of a library function
-  pub max_arity: Option<usize>,
-  /// The optimization strategy to use
-  pub strategy: OptimizationStrategy,
-}
-
-impl Default for ParetoConfig {
-  fn default() -> Self {
-    Self {
-      final_beams: 10,
-      inter_beams: 5,
-      lib_iter_limit: 3,
-      lps: 5,
-      learn_constants: false,
-      max_arity: None,
-      strategy: OptimizationStrategy::Balanced(0.5),
-    }
-  }
-}
-
-/// A BabbleRunner that uses Pareto-optimal beam search
-pub struct ParetoRunner<Op, LA, LD>
-where
-  Op: Display + Hash + Clone + Ord + Teachable + Arity + Send + Sync + 'static,
-  LA: LangCost<Op> + Clone + Send + Sync + 'static,
-  LD: LangGain<Op> + Clone + Send + Sync + 'static,
-{
-  /// The domain-specific rewrites to apply
-  dsrs: Vec<Rewrite<AstNode<Op>, BeamAreaDelay>>,
-  /// Configuration for the Pareto-optimal beam search
-  config: ParetoConfig,
-  /// The area cost model
-  area_cost: LA,
-  /// The delay cost model
-  delay_cost: LD,
-}
-
-impl<Op, LA, LD> ParetoRunner<Op, LA, LD>
-where
-  Op: Arity
-    + Teachable
-    + Printable
-    + Debug
-    + Default
-    + Display
-    + Hash
-    + Clone
-    + Ord
-    + Sync
-    + Send
-    + DiscriminantEq
-    + 'static,
-  LA: LangCost<Op> + Clone + Send + Sync + 'static,
-  LD: LangGain<Op> + Clone + Default + Send + Sync + 'static,
-{
-  /// Create a new ParetoRunner with the given domain-specific rewrites,
-  /// configuration, and cost models
-  pub fn new<I>(
-    dsrs: I,
-    config: ParetoConfig,
-    area_cost: LA,
-    delay_cost: LD,
-  ) -> Self
-  where
-    I: IntoIterator<Item = Rewrite<AstNode<Op>, BeamAreaDelay>>,
-  {
-    // 如果USE_RULES为false，将dsrs清空
-    let dsrs = if !USE_RULES {
-      Vec::new()
-    } else {
-      dsrs.into_iter().collect()
-    };
-    Self {
-      dsrs,
-      config,
-      area_cost,
-      delay_cost,
-    }
-  }
-
-  /// Run the e-graph and library learning process with Pareto-optimal beam
-  /// search
-  fn run_egraph(
-    &self,
-    roots: &[Id],
-    egraph: EGraph<AstNode<Op>, BeamAreaDelay>,
-  ) -> BabbleResult<Op> {
-    let start_time = Instant::now();
-    let timeout = Duration::from_secs(60 * 100_000);
-
-    info!("Initial egraph size: {}", egraph.total_size());
-    info!("Running {} DSRs... ", self.dsrs.len());
-
-    let runner = EggRunner::<_, _, ()>::new(BeamAreaDelay::empty())
-      .with_egraph(egraph)
-      .with_time_limit(timeout)
-      .with_iter_limit(3)
-      .run(&self.dsrs);
-
-    let aeg = runner.egraph;
-
-    info!(
-      "Finished in {}ms; final egraph size: {}",
-      start_time.elapsed().as_millis(),
-      aeg.total_size()
-    );
-
-    info!("Running co-occurrence analysis... ");
-    let co_time = Instant::now();
-    let co_ext = COBuilder::new(&aeg, roots);
-    let co_occurs = co_ext.run();
-    info!("Finished in {}ms", co_time.elapsed().as_millis());
-
-    info!("Running anti-unification... ");
-    let au_time = Instant::now();
-    let mut learned_lib = LearnedLibraryBuilder::default()
-      .learn_constants(self.config.learn_constants)
-      .max_arity(self.config.max_arity)
-      .with_co_occurs(co_occurs)
-      .build(&aeg, self.delay_cost.clone());
-    info!(
-      "Found {} patterns in {}ms",
-      learned_lib.size(),
-      au_time.elapsed().as_millis()
-    );
-
-    info!("Deduplicating patterns... ");
-    let dedup_time = Instant::now();
-    learned_lib.deduplicate(&aeg);
-    let lib_rewrites: Vec<_> = learned_lib.rewrites().collect();
-    info!(
-      "Reduced to {} patterns in {}ms",
-      learned_lib.size(),
-      dedup_time.elapsed().as_millis()
-    );
-
-    info!("Adding libs and running Pareto-optimal beam search... ");
-    let lib_rewrite_time = Instant::now();
-    let runner = EggRunner::<_, _, ()>::new(BeamAreaDelay::new(
-      self.config.final_beams,
-      self.config.inter_beams,
-      self.config.lps,
-      self.config.strategy,
-    ))
-    .with_egraph(aeg.clone())
-    .with_iter_limit(self.config.lib_iter_limit)
-    .with_time_limit(timeout)
-    .with_node_limit(1_000_000)
-    .run(lib_rewrites.iter());
-
-    let mut egraph = runner.egraph;
-    let root = egraph.add(AstNode::new(Op::list(), roots.iter().copied()));
-    let mut cs = egraph[egraph.find(root)].data.clone();
-
-    // Sort by combined cost according to the strategy
-    cs.set.sort_by(|a, b| {
-      a.combined_cost(&self.config.strategy)
-        .partial_cmp(&b.combined_cost(&self.config.strategy))
-        .unwrap_or(std::cmp::Ordering::Equal)
-    });
-
-    info!("Finished in {}ms", lib_rewrite_time.elapsed().as_millis());
-    info!("Stop reason: {:?}", runner.stop_reason.unwrap());
-    info!("Number of nodes: {}", egraph.total_size());
-
-    debug!("learned libs");
-    let all_libs: Vec<_> = learned_lib.libs().collect();
-    let mut chosen_rewrites = Vec::new();
-    for lib in &cs.set[0].libs {
-      debug!("{}: {}", lib.0, &all_libs[lib.0 .0]);
-      chosen_rewrites.push(lib_rewrites[lib.0 .0].clone());
-    }
-
-    debug!(
-      "area cost: {}, delay cost: {}",
-      cs.set[0].full_area_cost, cs.set[0].full_delay_cost
-    );
-
-    let ex_time = Instant::now();
-    info!("Extracting... ");
-    let lifted = apply_libs_area_delay(
-      aeg.clone(),
-      roots,
-      &chosen_rewrites,
-      self.area_cost.clone(),
-      self.delay_cost.clone(),
-      self.config.strategy,
-    );
-    let final_cost = AstSize.cost_rec(&lifted);
-
-    info!("Finished in {}ms", ex_time.elapsed().as_millis());
-    info!("final cost: {}", final_cost);
-    debug!("{}", Pretty::new(Arc::new(Expr::from(lifted.clone()))));
-    info!("round time: {}ms", start_time.elapsed().as_millis());
-
-    // Calculate initial cost
-    let initial_cost = {
-      let s: usize = roots
-        .iter()
-        .map(|id| {
-          let extractor = egg::Extractor::new(&egraph, AstSize);
-          let (_, expr) = extractor.find_best(*id);
-          AstSize.cost_rec(&expr)
-        })
-        .sum();
-      s + 1 // Add one to account for root node
-    };
-
-    // Convert BeamAreaDelay rewrites to PartialLibCost rewrites
-    let converted_rewrites: Vec<Rewrite<AstNode<Op>, PartialLibCost>> =
-      chosen_rewrites
-        .iter()
-        .map(|r| {
-          // Create a new empty rewrite with the same name
-          let empty_expr = RecExpr::<AstNode<Op>>::default();
-          let pattern = egg::Pattern::from(empty_expr.clone());
-          Rewrite::new(r.name.clone(), pattern, egg::Pattern::from(empty_expr))
-            .unwrap_or_else(|_| panic!("Failed to convert rewrite"))
-        })
-        .collect();
-
-    BabbleResult {
-      final_expr: lifted.into(),
-      num_libs: chosen_rewrites.len(),
-      rewrites: converted_rewrites,
-      initial_cost,
-      final_cost,
-      run_time: start_time.elapsed(),
-    }
-  }
-}
-
-impl<Op, LA, LD> BabbleRunner<Op> for ParetoRunner<Op, LA, LD>
-where
-  Op: Arity
-    + Teachable
-    + Printable
-    + Debug
-    + Default
-    + Display
-    + Hash
-    + Clone
-    + Ord
-    + Sync
-    + Send
-    + DiscriminantEq
-    + 'static,
-  LA: LangCost<Op> + Clone + Send + Sync + 'static,
-  LD: LangGain<Op> + Clone + Default + Send + Sync + 'static,
-{
-  fn run(&self, expr: Expr<Op>) -> BabbleResult<Op> {
-    self.run_multi(vec![expr])
-  }
-
-  fn run_multi(&self, exprs: Vec<Expr<Op>>) -> BabbleResult<Op> {
-    // First, let's turn our list of exprs into a list of recexprs
-    let recexprs: Vec<RecExpr<AstNode<Op>>> =
-      exprs.into_iter().map(RecExpr::from).collect();
-
-    let mut egraph = EGraph::new(BeamAreaDelay::new(
-      self.config.final_beams,
-      self.config.inter_beams,
-      self.config.lps,
-      self.config.strategy,
-    ));
-    let roots = recexprs
-      .iter()
-      .map(|x| egraph.add_expr(x))
-      .collect::<Vec<_>>();
-    egraph.rebuild();
-
-    self.run_egraph(&roots, egraph)
-  }
-
-  fn run_equiv_groups(
-    &self,
-    expr_groups: Vec<Vec<Expr<Op>>>,
-  ) -> BabbleResult<Op> {
-    // First, let's turn our list of exprs into a list of recexprs
-    let recexpr_groups: Vec<Vec<_>> = expr_groups
-      .into_iter()
-      .map(|group| group.into_iter().map(RecExpr::from).collect())
-      .collect();
-
-    let mut egraph = EGraph::new(BeamAreaDelay::new(
-      self.config.final_beams,
-      self.config.inter_beams,
-      self.config.lps,
-      self.config.strategy,
-    ));
-
-    let roots: Vec<_> = recexpr_groups
-      .into_iter()
-      .map(|mut group| {
-        let first_expr = group.pop().unwrap();
-        let root = egraph.add_expr(&first_expr);
-        for expr in group {
-          let class = egraph.add_expr(&expr);
-          egraph.union(root, class);
-        }
-        root
-      })
-      .collect();
-
-    egraph.rebuild();
-
-    self.run_egraph(&roots, egraph)
-  }
-}
-
-// Implement Debug that doesn't depend on Op being Debug
-impl<Op, LA, LD> std::fmt::Debug for ParetoRunner<Op, LA, LD>
-where
-  Op: Display + Hash + Clone + Ord + Teachable + Arity + Send + Sync + 'static,
-  LA: LangCost<Op> + Clone + Send + Sync + 'static,
-  LD: LangGain<Op> + Clone + Send + Sync + 'static,
-{
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    f.debug_struct("ParetoRunner")
-      .field("dsrs_count", &self.dsrs.len())
-      .field("config", &self.config)
-      .finish()
-  }
-}
-
-/// A trait for running library learning experiments with knapsack optimization
-pub trait BabbleKnapsackRunner<Op, LA, LD>
+/// A trait for running library learning experiments with Pareto optimization
+pub trait BabbleParetoRunner<Op, LA, LD>
 where
   Op: Arity
     + Teachable
@@ -774,21 +435,21 @@ where
   LD: LangGain<Op> + Clone + Default,
 {
   /// Run the experiment on a single expression
-  fn run(&self, expr: Expr<Op>) -> KnapsackResult<Op, LA, LD>;
+  fn run(&self, expr: Expr<Op>) -> ParetoResult<Op, LA, LD>;
 
   /// Run the experiment on multiple expressions
-  fn run_multi(&self, exprs: Vec<Expr<Op>>) -> KnapsackResult<Op, LA, LD>;
+  fn run_multi(&self, exprs: Vec<Expr<Op>>) -> ParetoResult<Op, LA, LD>;
 
   /// Run the experiment on groups of equivalent expressions
   fn run_equiv_groups(
     &self,
     expr_groups: Vec<Vec<Expr<Op>>>,
-  ) -> KnapsackResult<Op, LA, LD>;
+  ) -> ParetoResult<Op, LA, LD>;
 }
 
 /// Result of running a BabbleRunner experiment
 #[derive(Clone)]
-pub struct KnapsackResult<Op, LA, LD>
+pub struct ParetoResult<Op, LA, LD>
 where
   Op: Display + Hash + Clone + Ord + Teachable + Arity + Debug + 'static,
   LA: LangCost<Op> + Clone + Default,
@@ -801,33 +462,33 @@ where
   /// The rewrites representing the learned libraries
   pub rewrites: HashMap<
     usize,
-    Rewrite<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>,
+    Rewrite<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>,
   >,
-  /// The initial cost of the expression(s)
-  pub initial_cost: usize,
-  /// The final cost of the expression
-  pub final_cost: usize,
+  /// The initial cost of the expression
+  pub initial_cost: (usize, usize, f32),
+  /// The final cost of the expression (area_cost, delay_cost, balanced_cost)
+  pub final_cost: (usize, usize, f32),
   /// The time taken to run the experiment
   pub run_time: Duration,
 }
 
-impl<Op, LA, LD> KnapsackResult<Op, LA, LD>
+impl<Op, LA, LD> ParetoResult<Op, LA, LD>
 where
   Op: Display + Hash + Clone + Ord + Teachable + Arity + Debug + 'static,
   LA: LangCost<Op> + Clone + Default,
   LD: LangGain<Op> + Clone + Default,
 {
   /// Calculate the compression ratio achieved
-  pub fn compression_ratio(&self) -> f64 {
-    if self.initial_cost == 0 {
+  pub fn compression_ratio(&self) -> f32 {
+    if self.initial_cost.2 == 0.0 {
       1.0
     } else {
-      1.0 - (self.final_cost as f64 / self.initial_cost as f64)
+      1.0 - (self.final_cost.2 / self.initial_cost.2)
     }
   }
 }
 
-impl<Op, LA, LD> std::fmt::Debug for KnapsackResult<Op, LA, LD>
+impl<Op, LA, LD> std::fmt::Debug for ParetoResult<Op, LA, LD>
 where
   Op: Display + Hash + Clone + Ord + Teachable + Arity + Debug + 'static,
   LA: LangCost<Op> + Clone + Default,
@@ -836,8 +497,8 @@ where
   fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
     f.debug_struct("BabbleResult")
       .field("num_libs", &self.num_libs)
-      .field("initial_cost", &self.initial_cost)
-      .field("final_cost", &self.final_cost)
+      .field("initial_cost", &self.initial_cost.2)
+      .field("final_cost", &self.final_cost.2)
       .field("compression_ratio", &self.compression_ratio())
       .field("run_time", &self.run_time)
       .finish()
@@ -846,7 +507,7 @@ where
 
 /// Configuration for beam search
 #[derive(Debug, Clone, Copy)]
-pub struct KnapsackConfig {
+pub struct ParetoConfig {
   /// The final beam size to use
   pub final_beams: usize,
   /// The inter beam size to use
@@ -859,11 +520,11 @@ pub struct KnapsackConfig {
   pub learn_constants: bool,
   /// Maximum arity of a library function
   pub max_arity: Option<usize>,
-  /// Maximum library area cost
-  pub max_area: usize,
+  /// strategy of balancing area and delay
+  pub strategy: f32,
 }
 
-impl Default for KnapsackConfig {
+impl Default for ParetoConfig {
   fn default() -> Self {
     Self {
       final_beams: 10,
@@ -872,13 +533,13 @@ impl Default for KnapsackConfig {
       lps: 1,
       learn_constants: false,
       max_arity: None,
-      max_area: 100,
+      strategy: 0.8,
     }
   }
 }
 
 /// A BabbleRunner that uses regular beam search
-pub struct KnapsackRunner<Op, LA, LD>
+pub struct ParetoRunner<Op, LA, LD>
 where
   Op: Display
     + Hash
@@ -894,19 +555,19 @@ where
   LD: LangGain<Op> + Clone + Default,
 {
   /// The domain-specific rewrites to apply
-  dsrs: Vec<Rewrite<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>>,
+  dsrs: Vec<Rewrite<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>>,
   /// lib rewrites
   lib_rewrites: HashMap<
     usize,
-    Rewrite<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>,
+    Rewrite<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>,
   >,
   /// Configuration for the beam search
-  config: KnapsackConfig,
+  config: ParetoConfig,
   lang_cost: LA,
   lang_gain: LD,
 }
 
-impl<Op, LA, LD> KnapsackRunner<Op, LA, LD>
+impl<Op, LA, LD> ParetoRunner<Op, LA, LD>
 where
   Op: Arity
     + LibOperation
@@ -931,15 +592,15 @@ where
     dsrs: I,
     lib_rewrites: HashMap<
       usize,
-      Rewrite<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>,
+      Rewrite<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>,
     >,
-    config: KnapsackConfig,
+    config: ParetoConfig,
     lang_cost: LA,
     lang_gain: LD,
   ) -> Self
   where
     I: IntoIterator<
-      Item = Rewrite<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>,
+      Item = Rewrite<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>,
     >,
   {
     // 如果USE_RULES为false，将dsrs清空
@@ -961,26 +622,39 @@ where
   fn run_egraph(
     &self,
     roots: &[Id],
-    egraph: EGraph<AstNode<Op>, beam_knapsack::PartialLibCost<Op, LA, LD>>,
-  ) -> KnapsackResult<Op, LA, LD> {
+    egraph: EGraph<AstNode<Op>, beam_pareto::PartialLibCost<Op, LA, LD>>,
+  ) -> ParetoResult<Op, LA, LD> {
     let start_time = Instant::now();
     let timeout = Duration::from_secs(60 * 100_000);
 
-    let mut init_cost: usize = match roots.len() {
+    let mut init_delay: usize = match roots.len() {
       0 | 1 => 0,
       _ => 1,
     };
+    let mut init_area: usize = 0;
     for root in roots {
-      let extractor =
+      let delay_extractor =
         egg::Extractor::new(&egraph, DelayCost::new(self.lang_gain.clone()));
-      let (_, expr) = extractor.find_best(*root);
-      let cost = DelayCost::new(self.lang_gain.clone()).cost_rec(&expr);
-      let selected_cost = match cost.2 {
-        true => cost.0,
-        false => cost.1,
+      let (_, expr) = delay_extractor.find_best(*root);
+      let delay_cost = DelayCost::new(self.lang_gain.clone()).cost_rec(&expr);
+      let selected_delay_cost = match delay_cost.2 {
+        true => delay_cost.0,
+        false => delay_cost.1,
       };
-      init_cost += selected_cost;
+      init_delay += selected_delay_cost;
+
+      let area_extractor =
+        egg::Extractor::new(&egraph, AreaCost::new(self.lang_cost.clone()));
+      let (_, expr) = area_extractor.find_best(*root);
+      let area_cost = AreaCost::new(self.lang_cost.clone()).cost_rec(&expr);
+      let selected_area_cost = match area_cost.2 {
+        true => area_cost.0,
+        false => area_cost.1,
+      };
+      init_area += selected_area_cost;
     }
+    let init_cost = self.config.strategy * (init_delay as f32)
+      + (1.0 - self.config.strategy) * (init_area as f32);
 
     println!(
       "Initial egraph size: {}, eclasses: {}",
@@ -990,7 +664,7 @@ where
     println!("Running {} DSRs... ", self.dsrs.len());
 
     let runner =
-      EggRunner::<_, _, ()>::new(beam_knapsack::PartialLibCost::empty())
+      EggRunner::<_, _, ()>::new(beam_pareto::PartialLibCost::empty())
         .with_egraph(egraph)
         .with_time_limit(timeout)
         .with_iter_limit(3)
@@ -1059,7 +733,7 @@ where
     info!("Adding libs and running beam search... ");
     let lib_rewrite_time = Instant::now();
     let runner =
-      EggRunner::<_, _, ()>::new(beam_knapsack::PartialLibCost::new(
+      EggRunner::<_, _, ()>::new(beam_pareto::PartialLibCost::new(
         self.config.final_beams,
         self.config.inter_beams,
         self.config.lps,
@@ -1110,38 +784,49 @@ where
 
     let ex_time = Instant::now();
     info!("Extracting... ");
-    let best = apply_libs_knapsack(
+    let best = apply_libs_pareto(
       aeg.clone(),
       roots,
       &chosen_rewrites,
       self.lang_cost.clone(),
       self.lang_gain.clone(),
+      self.config.strategy,
     );
-    let final_cost = DelayCost::new(self.lang_gain.clone()).cost_rec(&best);
-    let selected_final_cost = match final_cost.2 {
-      true => final_cost.0,
-      false => final_cost.1,
+
+    let delay_cost = DelayCost::new(self.lang_gain.clone()).cost_rec(&best);
+    let selected_delay_cost = match delay_cost.2 {
+      true => delay_cost.0,
+      false => delay_cost.1,
     };
 
-    let lifted = extract::lift_libs(&best);
+    let area_cost = AreaCost::new(self.lang_cost.clone()).cost_rec(&best);
+    let selected_area_cost = match area_cost.2 {
+      true => area_cost.0,
+      false => area_cost.1,
+    };
+    let fin_cost = self.config.strategy * (selected_delay_cost as f32)
+      + (1.0 - self.config.strategy) * (selected_area_cost as f32);
+
+    // Lifting the lib will result in incorrect cost
+    // let lifted = extract::lift_libs(&best);
 
     info!("Finished in {}ms", ex_time.elapsed().as_millis());
-    info!("final cost: {}", selected_final_cost);
-    debug!("{}", Pretty::new(Arc::new(Expr::from(lifted.clone()))));
+    debug!("final cost: {}", fin_cost);
+    debug!("{}", Pretty::new(Arc::new(Expr::from(best.clone()))));
     info!("round time: {}ms", start_time.elapsed().as_millis());
 
-    KnapsackResult {
-      final_expr: lifted.into(),
+    ParetoResult {
+      final_expr: best.into(),
       num_libs: chosen_rewrites.len(),
       rewrites: rewrites_map,
-      initial_cost: init_cost,
-      final_cost: selected_final_cost,
+      initial_cost: (init_area, init_delay, init_cost),
+      final_cost: (selected_area_cost, selected_delay_cost, fin_cost),
       run_time: start_time.elapsed(),
     }
   }
 }
 
-impl<Op, LA, LD> BabbleKnapsackRunner<Op, LA, LD> for KnapsackRunner<Op, LA, LD>
+impl<Op, LA, LD> BabbleParetoRunner<Op, LA, LD> for ParetoRunner<Op, LA, LD>
 where
   Op: Arity
     + LibOperation
@@ -1160,16 +845,16 @@ where
   LA: LangCost<Op> + Clone + Default + 'static,
   LD: LangGain<Op> + Clone + Default + 'static,
 {
-  fn run(&self, expr: Expr<Op>) -> KnapsackResult<Op, LA, LD> {
+  fn run(&self, expr: Expr<Op>) -> ParetoResult<Op, LA, LD> {
     self.run_multi(vec![expr])
   }
 
-  fn run_multi(&self, exprs: Vec<Expr<Op>>) -> KnapsackResult<Op, LA, LD> {
+  fn run_multi(&self, exprs: Vec<Expr<Op>>) -> ParetoResult<Op, LA, LD> {
     // First, let's turn our list of exprs into a list of recexprs
     let recexprs: Vec<RecExpr<AstNode<Op>>> =
       exprs.into_iter().map(RecExpr::from).collect();
 
-    let mut egraph = EGraph::new(beam_knapsack::PartialLibCost::new(
+    let mut egraph = EGraph::new(beam_pareto::PartialLibCost::new(
       self.config.final_beams,
       self.config.inter_beams,
       self.config.lps,
@@ -1188,14 +873,14 @@ where
   fn run_equiv_groups(
     &self,
     expr_groups: Vec<Vec<Expr<Op>>>,
-  ) -> KnapsackResult<Op, LA, LD> {
+  ) -> ParetoResult<Op, LA, LD> {
     // First, let's turn our list of exprs into a list of recexprs
     let recexpr_groups: Vec<Vec<_>> = expr_groups
       .into_iter()
       .map(|group| group.into_iter().map(RecExpr::from).collect())
       .collect();
 
-    let mut egraph = EGraph::new(beam_knapsack::PartialLibCost::new(
+    let mut egraph = EGraph::new(beam_pareto::PartialLibCost::new(
       self.config.final_beams,
       self.config.inter_beams,
       self.config.lps,
@@ -1223,7 +908,7 @@ where
 }
 
 // Implement Debug that doesn't depend on Op being Debug
-impl<Op, LA, LD> std::fmt::Debug for KnapsackRunner<Op, LA, LD>
+impl<Op, LA, LD> std::fmt::Debug for ParetoRunner<Op, LA, LD>
 where
   Op: Display
     + Hash
